feat: reorganize views. Root Page is now in its own file.
